* Unified and disabled TRACEing for the usb stack and uhci hcd * Fixed warnings * Some cleanup * Added myself as an author to some files
No functional changes.

git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@18614 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/add-ons/kernel/bus_managers/usb/Transfer.cpp
@@ -3,6 +3,7 @@
  * Distributed under the terms of the MIT License.
  *
  * Authors:
+ *		Michael Lotz <mmlr@mlotz.ch>
  *		Niels S. Reedijk
  */
 
@@ -58,7 +59,7 @@ size_t
 Transfer::VectorLength()
 {
 	size_t length = 0;
-	for (int32 i = 0; i < fVectorCount; i++)
+	for (size_t i = 0; i < fVectorCount; i++)
 		length += fVector[i].iov_len;
 	return length;
 }
